# docs/swagger.yaml
definitions:
  article.CommentInput:
    properties:
      comment:
        type: string
    required:
    - comment
    type: object
  article.SaveArticleInput:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  auth.LoginInput:
    properties:
      password:
        type: string
      user_name:
        type: string
    required:
    - password
    - user_name
    type: object
  auth.RegisterInput:
    properties:
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      password_confirm:
        type: string
      user_name:
        type: string
    required:
    - full_name
    - password
    - password_confirm
    - user_name
    type: object
  following.FollowUserInput:
    properties:
      username:
        type: string
    type: object
  profile.ChangePasswordInput:
    properties:
      old_password:
        type: string
      password:
        minLength: 6
        type: string
      password_confirm:
        type: string
    required:
    - old_password
    - password
    - password_confirm
    type: object
info:
  contact: {}
paths:
  /articles:
    get:
      produces:
      - application/json
      responses: {}
      summary: Get random article.
      tags:
      - Articles
    post:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create article
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/article.SaveArticleInput'
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Create article.
      tags:
      - Articles
  /articles/{id}:
    delete:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Article id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Delete article.
      tags:
      - Articles
    put:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update article
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/article.SaveArticleInput'
      - description: Article id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Update article.
      tags:
      - Articles
  /articles/{username}:
    get:
      parameters:
      - description: Limit returning value
        in: query
        name: limit
        type: integer
      - description: Paging
        in: query
        name: page
        type: integer
      - description: user username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get user article.
      tags:
      - Articles
  /articles/{username}/{id}:
    get:
      parameters:
      - description: user username
        in: path
        name: username
        required: true
        type: string
      - description: user username
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Get detail article.
      tags:
      - Articles
  /articles/{username}/{id}/comment:
    post:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: user username
        in: path
        name: username
        required: true
        type: string
      - description: Article id
        in: path
        name: id
        required: true
        type: integer
      - description: the body to comment an article
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/article.CommentInput'
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Comment article.
      tags:
      - Articles
  /articles/{username}/{id}/like:
    post:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: user username
        in: path
        name: username
        required: true
        type: string
      - description: Article id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Likes article.
      tags:
      - Articles
  /auth/login:
    post:
      parameters:
      - description: the body to login
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses: {}
      summary: Login.
      tags:
      - Auth
  /auth/register:
    post:
      parameters:
      - description: the body to register
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterInput'
      produces:
      - application/json
      responses: {}
      summary: Register.
      tags:
      - Auth
  /followers:
    get:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Get follower user.
      tags:
      - Followers
  /following:
    get:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Get following user.
      tags:
      - Following
    post:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to follow a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/following.FollowUserInput'
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Follow a user.
      tags:
      - Following
  /following/{id}:
    delete:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Following id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Delete a following user.
      tags:
      - Following
  /profile:
    get:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      tags:
      - Profile
  /profile/change-password:
    post:
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to change password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/profile.ChangePasswordInput'
      produces:
      - application/json
      responses: {}
      security:
      - BearerToken: []
      summary: Change password user.
      tags:
      - Profile
swagger: "2.0"